refactor: re-structure the canary directory
