--- src/components/Canary/HealthSummary/index.ts
@@ -1,3 +1,0 @@
-import { HealthSummary } from "./summary";
-
-export { HealthSummary };
